Use ajax to submit the email survey form to feedback
Doing this means we don't interupt a user journey the user might
be on by taking them somewhere else just to show them a "thanks"
page.  We do need to add an error form though - decided that this
shouldn't set the cookie to stop them being shown the survey again
later.